refactor(backend): split the postgres integration harness into modules
run-postgres-integrations.mjs was 3,875 lines - the largest authored file in the repo -
because a full Postgres lifecycle supervisor had grown inside the test runner. Extract
boundaries, errors, connection, supervisedClient and identityGuards into
scripts/postgresIntegrations/, leaving a 2,645-line runner. Tooling only; the
boundaryDefinitions array moves unchanged.
