+++ apps/backend/scripts/postgresIntegrations/boundaries.mjs
@@ -1,0 +1,138 @@
+import { randomBytes } from "node:crypto";
+
+export const administrativeDatabaseName = "postgres";
+export const disposableDatabaseName = "flashcards";
+export const lifecycleLockKeys = Object.freeze([1196572995, 1886546277]);
+export const lifecycleCleanupMaximumAttempts = 2;
+export const lifecycleCleanupTimeoutMilliseconds = 5_000;
+export const lifecycleRecoveryPollMilliseconds = 100;
+export const mutableWorkShutdownPollMilliseconds = 50;
+export const databaseTerminationPollMilliseconds = 100;
+export const cleanupClientTeardownReserveMilliseconds = 1_000;
+export const emergencyClientTeardownReserveMilliseconds = 50;
+export const databaseOidMinimum = 16_384;
+export const databaseOidSelectionMaximumAttempts = 64;
+export const postgresConnectionTimeoutMilliseconds = 5_000;
+export const postgresStartupOptions =
+  "-c standard_conforming_strings=on -c client_encoding=UTF8";
+export const integrationChildDatabaseEnvironmentVariableNames = Object.freeze([
+  "DATABASE_URL",
+  "DB_AUTH_SECRET_ARN",
+  "DB_BACKEND_SECRET_ARN",
+  "DB_HOST",
+  "DB_NAME",
+  "DB_OWNER_SECRET_ARN",
+  "DB_REPORTING_SECRET_ARN",
+  "DB_SECRET_ARN",
+  "NODE_PG_FORCE_NATIVE",
+  "PGCLIENT_ENCODING",
+  "PGCLIENTENCODING",
+  "PGCONNECT_TIMEOUT",
+  "PGDATABASE",
+  "PGHOST",
+  "PGHOSTADDR",
+  "PGOPTIONS",
+  "PGPASSFILE",
+  "PGPASSWORD",
+  "PGPORT",
+  "PGREPLICATION",
+  "PGREQUIRESSL",
+  "PGSERVICE",
+  "PGSERVICEFILE",
+  "PGSSLCRL",
+  "PGSSLCRLDIR",
+  "PGSSLCERT",
+  "PGSSLKEY",
+  "PGSSLMODE",
+  "PGSSLNEGOTIATION",
+  "PGSSLROOTCERT",
+  "PGTARGETSESSIONATTRS",
+  "PGUSER",
+  "POSTGRES_INTEGRATION_ADMIN_URL",
+  "POSTGRES_INTEGRATION_EXPECTED_CLIENT_ENCODING",
+  "POSTGRES_INTEGRATION_EXPECTED_DATABASE_NAME",
+  "POSTGRES_INTEGRATION_EXPECTED_DATABASE_OID",
+  "POSTGRES_INTEGRATION_EXPECTED_OWNER_USERNAME",
+  "POSTGRES_INTEGRATION_EXPECTED_RUNTIME_USERNAME",
+  "REPORTING_DATABASE_URL",
+  "REPORTING_DB_SECRET_ARN",
+  "TEST_DATABASE_ADMIN_URL",
+]);
+export const managedRoleNames = Object.freeze([
+  "app",
+  "backend_app",
+  "auth_app",
+  "reporting_readonly",
+]);
+export const createdRolesByMigration = new Map([
+  ["0001_initial_schema.sql", Object.freeze(["app"])],
+  ["0024_auth_runtime_roles.sql", Object.freeze(["backend_app", "auth_app"])],
+  ["0044_reporting_readonly_role.sql", Object.freeze(["reporting_readonly"])],
+]);
+export const boundaryDefinitions = Object.freeze([
+  Object.freeze({
+    migrationFileName: "0104_generated_image_placeholder_terminal_state.sql",
+    expectedMigrationCount: 106,
+    testFiles: Object.freeze([
+      "src/cards/generatedImageAppend.postgres.integration.ts",
+      "src/chat/cardImages/operation.postgres.integration.ts",
+      "src/chat/cardImages/promotion/jobs.postgres.integration.ts",
+      "src/chat/runs/generatedImageAttemptBudget.postgres.integration.ts",
+      "src/database/aiChatInitiatingAuthClassification.postgres.integration.ts",
+      "src/mediaAssets/blobLifecycle/cleanup.postgres.integration.ts",
+    ]),
+  }),
+  Object.freeze({
+    migrationFileName: "0101_multipart_foreground_completion_fencing.sql",
+    expectedMigrationCount: 103,
+    testFiles: Object.freeze([
+      "src/mediaAssets/multipart/atomicWriter.postgres.integration.ts",
+      "src/mediaAssets/multipart/foregroundFencing.postgres.integration.ts",
+    ]),
+  }),
+  Object.freeze({
+    migrationFileName: "0100_multipart_replacement_creation_claim.sql",
+    expectedMigrationCount: 102,
+    testFiles: Object.freeze([
+      "src/mediaAssets/multipart/replacementCreationClaim.postgres.integration.ts",
+      "src/mediaAssets/multipart/uploadSessionCreation.postgres.integration.ts",
+    ]),
+  }),
+  Object.freeze({
+    migrationFileName: "0099_durable_multipart_completion_reconciliation.sql",
+    expectedMigrationCount: 101,
+    testFiles: Object.freeze([
+      "src/database/deadline.postgres.integration.ts",
+      "src/mediaAssets/blobLifecycle/lifecycle.postgres.integration.ts",
+      "src/mediaAssets/ingestion/directIngestionApply.postgres.integration.ts",
+      "src/mediaAssets/multipart/completionReconciliation.postgres.integration.ts",
+      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
+    ]),
+  }),
+  Object.freeze({
+    migrationFileName: "0098_multipart_writer_abort_and_terminal_replay.sql",
+    expectedMigrationCount: 100,
+    testFiles: Object.freeze([
+      "src/database/deadline.postgres.integration.ts",
+      "src/mediaAssets/blobLifecycle/lifecycle.postgres.integration.ts",
+      "src/mediaAssets/ingestion/directIngestionApply.postgres.integration.ts",
+      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
+    ]),
+  }),
+  Object.freeze({
+    migrationFileName: "0097_direct_multipart_writer_attempt_fencing.sql",
+    expectedMigrationCount: 99,
+    testFiles: Object.freeze([
+      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
+    ]),
+  }),
+  Object.freeze({
+    migrationFileName: "0096_atomic_multipart_completion_resolution.sql",
+    expectedMigrationCount: 98,
+    testFiles: Object.freeze([
+      "src/mediaAssets/multipart/writerAttempts.postgres.integration.ts",
+    ]),
+  }),
+]);
+export const backendRolePassword =
+  `postgres-integration-${randomBytes(18).toString("hex")}`;
