+++ apps/backend/scripts/run-postgres-integrations.mjs
@@ -5,8 +5,61 @@ import { dirname, join, resolve } from "node:path";
 import { performance } from "node:perf_hooks";
 import { setTimeout as delay } from "node:timers/promises";
 import { fileURLToPath, pathToFileURL } from "node:url";
-import PgClient from "pg/lib/client.js";
 import pgUtils from "pg/lib/utils.js";
+import {
+  backendRolePassword,
+  boundaryDefinitions,
+  cleanupClientTeardownReserveMilliseconds,
+  createdRolesByMigration,
+  databaseOidMinimum,
+  databaseOidSelectionMaximumAttempts,
+  databaseTerminationPollMilliseconds,
+  disposableDatabaseName,
+  emergencyClientTeardownReserveMilliseconds,
+  lifecycleCleanupMaximumAttempts,
+  lifecycleLockKeys,
+  managedRoleNames,
+  postgresConnectionTimeoutMilliseconds,
+} from "./postgresIntegrations/boundaries.mjs";
+import {
+  asError,
+  cleanupOperationTimeoutError,
+  cleanupRemainingMilliseconds,
+  combineErrors,
+  contextualError,
+  createTerminalCleanupController,
+} from "./postgresIntegrations/errors.mjs";
+import {
+  createDatabaseUrl,
+  createIntegrationChildEnvironment,
+  createPostgresClientOptions,
+  requireAdminDatabaseConnection,
+} from "./postgresIntegrations/connection.mjs";
+import {
+  assertPostgresSessionContractState,
+  createMutableWorkSupervisor,
+  createSupervisedPostgresClient,
+  postgresSessionContractSql,
+  requireCleanupSessionContract,
+  requirePostgresSessionContract,
+  runCleanupClientOperation,
+  runCleanupQuery,
+  runEmergencyClientOperation,
+  withPostgresClient,
+} from "./postgresIntegrations/supervisedClient.mjs";
+import {
+  acquireLifecycleLock,
+  cleanupIdentityConflictErrors,
+  databaseIdentityInterferenceError,
+  getDatabaseOwnershipEvidence,
+  inspectBoundaryCleanupIdentities,
+  inspectDatabaseIdentity,
+  inspectDatabaseIdentityForCleanup,
+  inspectOwnedRoleIdentities,
+  reacquireLifecycleLockForCleanup,
+  roleIdentityInterferenceError,
+  terminateDatabaseSessionsByOid,
+} from "./postgresIntegrations/identityGuards.mjs";
 
 const { escapeIdentifier, escapeLiteral } = pgUtils;
 
@@ -14,1289 +67,6 @@ const scriptDirectory = dirname(fileURLToPath(import.meta.url));
 const backendRoot = dirname(scriptDirectory);
 const repositoryRoot = resolve(backendRoot, "..", "..");
 const migrationsDirectory = join(repositoryRoot, "db", "migrations");
-const administrativeDatabaseName = "postgres";
-const disposableDatabaseName = "flashcards";
-const lifecycleLockKeys = Object.freeze([1196572995, 1886546277]);
-const lifecycleCleanupMaximumAttempts = 2;
-const lifecycleCleanupTimeoutMilliseconds = 5_000;
-const lifecycleRecoveryPollMilliseconds = 100;
-const mutableWorkShutdownPollMilliseconds = 50;
-const databaseTerminationPollMilliseconds = 100;
-const cleanupClientTeardownReserveMilliseconds = 1_000;
-const emergencyClientTeardownReserveMilliseconds = 50;
-const databaseOidMinimum = 16_384;
-const databaseOidSelectionMaximumAttempts = 64;
-const postgresConnectionTimeoutMilliseconds = 5_000;
-const postgresStartupOptions =
-  "-c standard_conforming_strings=on -c client_encoding=UTF8";
-const integrationChildDatabaseEnvironmentVariableNames = Object.freeze([
-  "DATABASE_URL",
-  "DB_AUTH_SECRET_ARN",
-  "DB_BACKEND_SECRET_ARN",
-  "DB_HOST",
-  "DB_NAME",
-  "DB_OWNER_SECRET_ARN",
-  "DB_REPORTING_SECRET_ARN",
-  "DB_SECRET_ARN",
-  "NODE_PG_FORCE_NATIVE",
-  "PGCLIENT_ENCODING",
-  "PGCLIENTENCODING",
-  "PGCONNECT_TIMEOUT",
-  "PGDATABASE",
-  "PGHOST",
-  "PGHOSTADDR",
-  "PGOPTIONS",
-  "PGPASSFILE",
-  "PGPASSWORD",
-  "PGPORT",
-  "PGREPLICATION",
-  "PGREQUIRESSL",
-  "PGSERVICE",
-  "PGSERVICEFILE",
-  "PGSSLCRL",
-  "PGSSLCRLDIR",
-  "PGSSLCERT",
-  "PGSSLKEY",
-  "PGSSLMODE",
-  "PGSSLNEGOTIATION",
-  "PGSSLROOTCERT",
-  "PGTARGETSESSIONATTRS",
-  "PGUSER",
-  "POSTGRES_INTEGRATION_ADMIN_URL",
-  "POSTGRES_INTEGRATION_EXPECTED_CLIENT_ENCODING",
-  "POSTGRES_INTEGRATION_EXPECTED_DATABASE_NAME",
-  "POSTGRES_INTEGRATION_EXPECTED_DATABASE_OID",
-  "POSTGRES_INTEGRATION_EXPECTED_OWNER_USERNAME",
-  "POSTGRES_INTEGRATION_EXPECTED_RUNTIME_USERNAME",
-  "REPORTING_DATABASE_URL",
-  "REPORTING_DB_SECRET_ARN",
-  "TEST_DATABASE_ADMIN_URL",
-]);
-const managedRoleNames = Object.freeze([
-  "app",
-  "backend_app",
-  "auth_app",
-  "reporting_readonly",
-]);
-const createdRolesByMigration = new Map([
-  ["0001_initial_schema.sql", Object.freeze(["app"])],
-  ["0024_auth_runtime_roles.sql", Object.freeze(["backend_app", "auth_app"])],
-  ["0044_reporting_readonly_role.sql", Object.freeze(["reporting_readonly"])],
-]);
-const boundaryDefinitions = Object.freeze([
-  Object.freeze({
-    migrationFileName: "0104_generated_image_placeholder_terminal_state.sql",
-    expectedMigrationCount: 106,
-    testFiles: Object.freeze([
-      "src/cards/generatedImageAppend.postgres.integration.ts",
-      "src/chat/cardImages/operation.postgres.integration.ts",
-      "src/chat/cardImages/promotion/jobs.postgres.integration.ts",
-      "src/chat/runs/generatedImageAttemptBudget.postgres.integration.ts",
-      "src/database/aiChatInitiatingAuthClassification.postgres.integration.ts",
-      "src/mediaAssets/blobLifecycle/cleanup.postgres.integration.ts",
-    ]),
-  }),
-  Object.freeze({
-    migrationFileName: "0101_multipart_foreground_completion_fencing.sql",
-    expectedMigrationCount: 103,
-    testFiles: Object.freeze([
-      "src/mediaAssets/multipart/atomicWriter.postgres.integration.ts",
-      "src/mediaAssets/multipart/foregroundFencing.postgres.integration.ts",
-    ]),
-  }),
-  Object.freeze({
-    migrationFileName: "0100_multipart_replacement_creation_claim.sql",
-    expectedMigrationCount: 102,
-    testFiles: Object.freeze([
-      "src/mediaAssets/multipart/replacementCreationClaim.postgres.integration.ts",
-      "src/mediaAssets/multipart/uploadSessionCreation.postgres.integration.ts",
-    ]),
-  }),
-  Object.freeze({
-    migrationFileName: "0099_durable_multipart_completion_reconciliation.sql",
-    expectedMigrationCount: 101,
-    testFiles: Object.freeze([
-      "src/database/deadline.postgres.integration.ts",
-      "src/mediaAssets/blobLifecycle/lifecycle.postgres.integration.ts",
-      "src/mediaAssets/ingestion/directIngestionApply.postgres.integration.ts",
-      "src/mediaAssets/multipart/completionReconciliation.postgres.integration.ts",
-      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
-    ]),
-  }),
-  Object.freeze({
-    migrationFileName: "0098_multipart_writer_abort_and_terminal_replay.sql",
-    expectedMigrationCount: 100,
-    testFiles: Object.freeze([
-      "src/database/deadline.postgres.integration.ts",
-      "src/mediaAssets/blobLifecycle/lifecycle.postgres.integration.ts",
-      "src/mediaAssets/ingestion/directIngestionApply.postgres.integration.ts",
-      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
-    ]),
-  }),
-  Object.freeze({
-    migrationFileName: "0097_direct_multipart_writer_attempt_fencing.sql",
-    expectedMigrationCount: 99,
-    testFiles: Object.freeze([
-      "src/mediaAssets/multipart/writerAbortReplay.postgres.integration.ts",
-    ]),
-  }),
-  Object.freeze({
-    migrationFileName: "0096_atomic_multipart_completion_resolution.sql",
-    expectedMigrationCount: 98,
-    testFiles: Object.freeze([
-      "src/mediaAssets/multipart/writerAttempts.postgres.integration.ts",
-    ]),
-  }),
-]);
-const backendRolePassword =
-  `postgres-integration-${randomBytes(18).toString("hex")}`;
-
-function asError(error) {
-  return error instanceof Error ? error : new Error(String(error));
-}
-
-function contextualError(message, error) {
-  const cause = asError(error);
-  return new Error(`${message} cause=${cause.message}`, { cause });
-}
-
-function rootError(error) {
-  let current = asError(error);
-  const visited = new Set();
-  while (
-    current.cause instanceof Error
-    && !visited.has(current.cause)
-  ) {
-    visited.add(current);
-    current = current.cause;
-  }
-  return current;
-}
-
-function combineErrors(errors, message) {
-  const seenRootErrors = new Set();
-  const uniqueErrors = errors.filter((error) => {
-    const root = rootError(error);
-    if (seenRootErrors.has(root)) return false;
-    seenRootErrors.add(root);
-    return true;
-  });
-  if (uniqueErrors.length === 0) return null;
-  if (uniqueErrors.length === 1) return uniqueErrors[0];
-  return new AggregateError(uniqueErrors, message);
-}
-
-function createCleanupDeadline(phase) {
-  return Object.freeze({
-    expiresAt: performance.now() + lifecycleCleanupTimeoutMilliseconds,
-    phase,
-    timeoutMilliseconds: lifecycleCleanupTimeoutMilliseconds,
-  });
-}
-
-function createTerminalCleanupController() {
-  let activeDeadline = null;
-  return Object.freeze({
-    begin(phase) {
-      if (activeDeadline === null) {
-        activeDeadline = createCleanupDeadline(phase);
-      }
-      return activeDeadline;
-    },
-    clearCompletedBoundary(cleanupDeadline) {
-      if (activeDeadline !== cleanupDeadline) {
-        throw new Error(
-          "PostgreSQL integration terminal cleanup deadline changed unexpectedly while completing a boundary.",
-        );
-      }
-      activeDeadline = null;
-    },
-  });
-}
-
-function cleanupRemainingMilliseconds(cleanupDeadline, phase) {
-  const remainingMilliseconds = Math.ceil(
-    cleanupDeadline.expiresAt - performance.now(),
-  );
-  if (remainingMilliseconds <= 0) {
-    const error = new Error(
-      `PostgreSQL integration cleanup deadline expired before starting an operation. cleanupPhase=${cleanupDeadline.phase} operationPhase=${phase} timeoutMilliseconds=${cleanupDeadline.timeoutMilliseconds}`,
-    );
-    error.name = "PostgresIntegrationCleanupTimeoutError";
-    throw error;
-  }
-  return remainingMilliseconds;
-}
-
-function cleanupOperationTimeoutError(cleanupDeadline, phase) {
-  const error = new Error(
-    `PostgreSQL integration cleanup operation exceeded the shared deadline and its client connection was forcibly terminated. cleanupPhase=${cleanupDeadline.phase} operationPhase=${phase} timeoutMilliseconds=${cleanupDeadline.timeoutMilliseconds}`,
-  );
-  error.name = "PostgresIntegrationCleanupTimeoutError";
-  return error;
-}
-
-function decodeUrlAuthorityComponent(value, componentName) {
-  try {
-    return decodeURIComponent(value);
-  } catch (error) {
-    throw contextualError(
-      `POSTGRES_INTEGRATION_ADMIN_URL contains an invalid percent-encoded authority component. component=${componentName}`,
-      error,
-    );
-  }
-}
-
-function assertDisabledSslQuery(url, sourceName) {
-  const parameterNames = [...new Set(url.searchParams.keys())];
-  const unsupportedParameterNames = parameterNames
-    .filter((parameterName) => parameterName !== "sslmode");
-  if (unsupportedParameterNames.length > 0) {
-    throw new Error(
-      `${sourceName} contains unsupported PostgreSQL query parameters. supportedParameters=["sslmode"] actualUnsupportedParameters=${JSON.stringify(unsupportedParameterNames)}`,
-    );
-  }
-  const sslModes = url.searchParams.getAll("sslmode");
-  if (sslModes.length !== 1 || sslModes[0] !== "disable") {
-    throw new Error(
-      `${sourceName} must contain exactly one explicit sslmode=disable query parameter for the isolated local PostgreSQL service.`,
-    );
-  }
-}
-
-function requireAdminDatabaseConnection() {
-  const value = process.env.POSTGRES_INTEGRATION_ADMIN_URL?.trim();
-  if (value === undefined || value === "") {
-    throw new Error(
-      `POSTGRES_INTEGRATION_ADMIN_URL is required and must target the isolated PostgreSQL 18 ${administrativeDatabaseName} administrative database.`,
-    );
-  }
-
-  let url;
-  try {
-    url = new URL(value);
-  } catch {
-    throw new Error(
-      "POSTGRES_INTEGRATION_ADMIN_URL is not a valid URL. Provide a PostgreSQL URL with percent-encoded authority credentials, an explicit IPv4 or DNS hostname and port, the postgres database, and sslmode=disable.",
-    );
-  }
-  if (url.protocol !== "postgres:" && url.protocol !== "postgresql:") {
-    throw new Error(
-      "POSTGRES_INTEGRATION_ADMIN_URL must use the postgres or postgresql URL scheme.",
-    );
-  }
-  if (url.hostname === "") {
-    throw new Error(
-      "POSTGRES_INTEGRATION_ADMIN_URL must include an explicit hostname.",
-    );
-  }
-  if (url.hostname.startsWith("[") || url.hostname.endsWith("]")) {
-    throw new Error(
-      "POSTGRES_INTEGRATION_ADMIN_URL does not support IPv6 literal hosts because the installed PostgreSQL child URL parser does not preserve parent/child host parity. Use an IPv4 address or DNS hostname.",
-    );
-  }
-  if (url.port === "") {
-    throw new Error(
-      "POSTGRES_INTEGRATION_ADMIN_URL must include an explicit numeric port.",
-    );
-  }
-  const port = Number(url.port);
-  if (!Number.isInteger(port) || port < 1 || port > 65_535) {
-    throw new Error(
-      `POSTGRES_INTEGRATION_ADMIN_URL contains an invalid port. port=${url.port}`,
-    );
-  }
-  const urlDatabaseName = url.pathname.replace(/^\/+/u, "");
-  if (urlDatabaseName !== administrativeDatabaseName) {
-    throw new Error(
-      `POSTGRES_INTEGRATION_ADMIN_URL must target the canonical ${administrativeDatabaseName} administrative database so every runner instance uses the same lifecycle lock. actualDatabase=${urlDatabaseName || "(empty)"}`,
-    );
-  }
-  if (url.username === "" || url.password === "") {
-    throw new Error(
-      "POSTGRES_INTEGRATION_ADMIN_URL must include administrative username and password credentials in the URL authority.",
-    );
-  }
-  if (url.hash !== "") {
-    throw new Error(
-      "POSTGRES_INTEGRATION_ADMIN_URL must not contain a URL fragment.",
-    );
-  }
-  assertDisabledSslQuery(url, "POSTGRES_INTEGRATION_ADMIN_URL");
-  const username = decodeUrlAuthorityComponent(url.username, "username");
-  const password = decodeUrlAuthorityComponent(url.password, "password");
-  return Object.freeze({
-    database: administrativeDatabaseName,
-    host: url.hostname,
-    password,
-    port,
-    ssl: false,
-    username,
-  });
-}
-
-function createDatabaseUrl(
-  adminConnection,
-  databaseName,
-  username,
-  password,
-) {
-  const url = new URL("postgresql://localhost");
-  url.hostname = adminConnection.host.includes(":")
-    ? `[${adminConnection.host}]`
-    : adminConnection.host;
-  url.port = String(adminConnection.port);
-  url.pathname = `/${databaseName}`;
-  url.username = encodeURIComponent(username);
-  url.password = encodeURIComponent(password);
-  url.searchParams.set("sslmode", "disable");
-  return url.toString();
-}
-
-function createPostgresClientOptions(
-  adminConnection,
-  databaseName,
-  username,
-  password,
-  applicationName,
-) {
-  return {
-    application_name: applicationName,
-    client_encoding: "UTF8",
-    connectionTimeoutMillis: postgresConnectionTimeoutMilliseconds,
-    database: databaseName,
-    host: adminConnection.host,
-    options: postgresStartupOptions,
-    password,
-    port: adminConnection.port,
-    replication: "false",
-    ssl: false,
-    sslnegotiation: "postgres",
-    user: username,
-  };
-}
-
-function assertDisposableDatabaseUrl(databaseUrl, environmentVariableName) {
-  const url = new URL(databaseUrl);
-  if (url.protocol !== "postgresql:") {
-    throw new Error(
-      `PostgreSQL integration child URL must use the canonical postgresql scheme. environmentVariable=${environmentVariableName} actualProtocol=${url.protocol}`,
-    );
-  }
-  if (url.hostname === "" || url.port === "") {
-    throw new Error(
-      `PostgreSQL integration child URL must contain an explicit host and port. environmentVariable=${environmentVariableName}`,
-    );
-  }
-  const databaseName = url.pathname.replace(/^\/+/u, "");
-  if (databaseName !== disposableDatabaseName) {
-    throw new Error(
-      `PostgreSQL integration child URL must target the disposable database. environmentVariable=${environmentVariableName} expectedDatabase=${disposableDatabaseName} actualDatabase=${databaseName || "(empty)"}`,
-    );
-  }
-  if (url.username === "" || url.password === "") {
-    throw new Error(
-      `PostgreSQL integration child URL must contain authority credentials. environmentVariable=${environmentVariableName}`,
-    );
-  }
-  if (url.hash !== "") {
-    throw new Error(
-      `PostgreSQL integration child URL must not contain a URL fragment. environmentVariable=${environmentVariableName}`,
-    );
-  }
-  assertDisabledSslQuery(
-    url,
-    `PostgreSQL integration child URL environmentVariable=${environmentVariableName}`,
-  );
-  return url;
-}
-
-function createIntegrationChildEnvironment(
-  parentEnvironment,
-  runtimeDatabaseUrl,
-  ownerDatabaseUrl,
-  expectedDatabaseOid,
-) {
-  const runtimeUrl = assertDisposableDatabaseUrl(
-    runtimeDatabaseUrl,
-    "DATABASE_URL",
-  );
-  const ownerUrl = assertDisposableDatabaseUrl(
-    ownerDatabaseUrl,
-    "TEST_DATABASE_ADMIN_URL",
-  );
-  const runtimeUsername = decodeURIComponent(runtimeUrl.username);
-  const ownerUsername = decodeURIComponent(ownerUrl.username);
-  if (runtimeUsername !== "backend_app") {
-    throw new Error(
-      `PostgreSQL integration runtime URL must authenticate as backend_app. actualUsername=${runtimeUsername}`,
-    );
-  }
-  if (
-    runtimeUrl.protocol !== ownerUrl.protocol
-    || runtimeUrl.host !== ownerUrl.host
-  ) {
-    throw new Error(
-      "PostgreSQL integration runtime and owner URLs must target the same PostgreSQL server.",
-    );
-  }
-  if (!/^[0-9]+$/u.test(expectedDatabaseOid)) {
-    throw new Error(
-      `PostgreSQL integration child environment requires an exact numeric database OID. actualOid=${expectedDatabaseOid}`,
-    );
-  }
-
-  const sanitizedEnvironment = Object.fromEntries(
-    Object.entries(parentEnvironment)
-      .filter(([variableName]) => (
-        !integrationChildDatabaseEnvironmentVariableNames.includes(variableName)
-      )),
-  );
-  return {
-    ...sanitizedEnvironment,
-    DATABASE_URL: runtimeDatabaseUrl,
-    PGCLIENT_ENCODING: "UTF8",
-    PGOPTIONS: postgresStartupOptions,
-    POSTGRES_INTEGRATION_EXPECTED_CLIENT_ENCODING: "UTF8",
-    POSTGRES_INTEGRATION_EXPECTED_DATABASE_NAME: disposableDatabaseName,
-    POSTGRES_INTEGRATION_EXPECTED_DATABASE_OID: expectedDatabaseOid,
-    POSTGRES_INTEGRATION_EXPECTED_OWNER_USERNAME: ownerUsername,
-    POSTGRES_INTEGRATION_EXPECTED_RUNTIME_USERNAME: runtimeUsername,
-    TEST_DATABASE_ADMIN_URL: ownerDatabaseUrl,
-  };
-}
-
-const postgresSessionContractSql = `
-  SELECT
-    current_database() AS database_name,
-    current_setting('client_encoding') AS client_encoding,
-    current_setting('standard_conforming_strings') AS standard_conforming_strings
-`;
-
-function assertPostgresSessionContractState(
-  state,
-  expectedDatabaseName,
-  phase,
-) {
-  if (
-    state?.database_name !== expectedDatabaseName
-    || state.client_encoding !== "UTF8"
-    || state.standard_conforming_strings !== "on"
-  ) {
-    throw new Error(
-      `PostgreSQL session contract is invalid. phase=${phase} expectedDatabase=${expectedDatabaseName} actualDatabase=${state?.database_name ?? "missing"} expectedClientEncoding=UTF8 actualClientEncoding=${state?.client_encoding ?? "missing"} expectedStandardConformingStrings=on actualStandardConformingStrings=${state?.standard_conforming_strings ?? "missing"}`,
-    );
-  }
-}
-
-async function requirePostgresSessionContract(
-  client,
-  expectedDatabaseName,
-  phase,
-) {
-  const result = await client.query(postgresSessionContractSql);
-  assertPostgresSessionContractState(
-    result.rows[0],
-    expectedDatabaseName,
-    phase,
-  );
-}
-
-function createSupervisedPostgresClient(
-  clientOptions,
-  phase,
-  onBackgroundFailure,
-) {
-  const client = new PgClient(clientOptions);
-  let backgroundFailure = null;
-  let closePromise = null;
-  const handleError = (error) => {
-    if (backgroundFailure !== null) return;
-    backgroundFailure = contextualError(
-      `PostgreSQL client emitted a background connection error. phase=${phase}`,
-      error,
-    );
-    if (onBackgroundFailure !== null) {
-      onBackgroundFailure(backgroundFailure);
-    }
-  };
-  client.on("error", handleError);
-  const beginClose = () => {
-    if (closePromise !== null) return closePromise;
-    try {
-      closePromise = Promise.resolve(client.end());
-    } catch (error) {
-      closePromise = Promise.reject(error);
-    }
-    return closePromise;
-  };
-  return Object.freeze({
-    abort() {
-      const closing = beginClose();
-      client.connection.stream.destroy();
-      return closing;
-    },
-    backgroundFailure() {
-      return backgroundFailure;
-    },
-    client,
-    close() {
-      return beginClose();
-    },
-    detach() {
-      client.off("error", handleError);
-    },
-  });
-}
-
-async function finishTimedOutPostgresOperation(
-  cleanupDeadline,
-  phase,
-  timeoutError,
-  terminateTimedOutOperation,
-) {
-  const teardownPromise = Promise.resolve()
-    .then(terminateTimedOutOperation);
-  const teardownMilliseconds = Math.max(
-    0,
-    Math.floor(cleanupDeadline.expiresAt - performance.now()),
-  );
-  let teardownTimeoutHandle;
-  const teardownTimeout = new Promise((resolve, reject) => {
-    teardownTimeoutHandle = setTimeout(() => {
-      reject(new Error(
-        `Timed out awaiting forced PostgreSQL client/backend teardown. cleanupPhase=${cleanupDeadline.phase} operationPhase=${phase} timeoutMilliseconds=${cleanupDeadline.timeoutMilliseconds}`,
-      ));
-    }, teardownMilliseconds);
-  });
-  try {
-    await Promise.race([teardownPromise, teardownTimeout]);
-  } catch (teardownError) {
-    throw new AggregateError(
-      [timeoutError, asError(teardownError)],
-      `PostgreSQL operation and forced client/backend teardown exceeded the shared deadline. cleanupPhase=${cleanupDeadline.phase} operationPhase=${phase}`,
-    );
-  } finally {
-    clearTimeout(teardownTimeoutHandle);
-  }
-  throw timeoutError;
-}
-
-async function runDeadlineBoundPostgresOperation(
-  cleanupDeadline,
-  phase,
-  teardownReserveMilliseconds,
-  operation,
-  terminateTimedOutOperation,
-) {
-  const remainingMilliseconds = cleanupRemainingMilliseconds(
-    cleanupDeadline,
-    phase,
-  );
-  if (remainingMilliseconds <= teardownReserveMilliseconds) {
-    await finishTimedOutPostgresOperation(
-      cleanupDeadline,
-      phase,
-      cleanupOperationTimeoutError(cleanupDeadline, phase),
-      terminateTimedOutOperation,
-    );
-  }
-  const operationTimeoutMilliseconds =
-    remainingMilliseconds - teardownReserveMilliseconds;
-  const neverSettles = new Promise(() => {});
-  let timedOut = false;
-  let timeoutHandle;
-  const operationPromise = Promise.resolve()
-    .then(operation)
-    .then(
-      (result) => (timedOut ? neverSettles : result),
-      (error) => {
-        if (timedOut) return neverSettles;
-        throw error;
-      },
-    );
-  const timeoutPromise = new Promise((resolve, reject) => {
-    timeoutHandle = setTimeout(() => {
-      timedOut = true;
-      const timeoutError = cleanupOperationTimeoutError(
-        cleanupDeadline,
-        phase,
-      );
-      finishTimedOutPostgresOperation(
-        cleanupDeadline,
-        phase,
-        timeoutError,
-        terminateTimedOutOperation,
-      ).catch(reject);
-    }, operationTimeoutMilliseconds);
-  });
-  try {
-    return await Promise.race([operationPromise, timeoutPromise]);
-  } finally {
-    clearTimeout(timeoutHandle);
-  }
-}
-
-function runCleanupClientOperation(
-  cleanupSession,
-  cleanupDeadline,
-  phase,
-  operation,
-) {
-  return runDeadlineBoundPostgresOperation(
-    cleanupDeadline,
-    phase,
-    cleanupClientTeardownReserveMilliseconds,
-    operation,
-    () => {
-      cleanupSession.connected = false;
-      cleanupSession.lockHeld = false;
-      return cleanupSession.terminateTimedOutOperation(
-        cleanupDeadline,
-        phase,
-      );
-    },
-  );
-}
-
-async function runEmergencyClientOperation(
-  supervisedClient,
-  cleanupDeadline,
-  phase,
-  operation,
-) {
-  return runDeadlineBoundPostgresOperation(
-    cleanupDeadline,
-    phase,
-    emergencyClientTeardownReserveMilliseconds,
-    operation,
-    () => supervisedClient.abort(),
-  );
-}
-
-function runCleanupQuery(
-  cleanupSession,
-  cleanupDeadline,
-  phase,
-  text,
-  values,
-) {
-  return runCleanupClientOperation(
-    cleanupSession,
-    cleanupDeadline,
-    phase,
-    () => cleanupSession.supervisedClient.client.query(text, values),
-  );
-}
-
-async function requireCleanupSessionContract(
-  cleanupSession,
-  cleanupDeadline,
-  expectedDatabaseName,
-  phase,
-) {
-  const result = await runCleanupQuery(
-    cleanupSession,
-    cleanupDeadline,
-    phase,
-    postgresSessionContractSql,
-    [],
-  );
-  assertPostgresSessionContractState(
-    result.rows[0],
-    expectedDatabaseName,
-    phase,
-  );
-}
-
-function createMutableWorkSupervisor() {
-  let firstShutdownFailure = null;
-  let nextWorkId = 1;
-  const activeWork = new Map();
-  const shutdownErrors = [];
-  let resolveShutdown;
-  const shutdown = new Promise((resolveStartedShutdown) => {
-    resolveShutdown = resolveStartedShutdown;
-  });
-
-  const finishEntry = (entry) => {
-    if (
-      !entry.completed
-      || entry.rawOperations.size > 0
-      || (entry.stopPromise !== null && !entry.stopSettled)
-    ) {
-      return;
-    }
-    activeWork.delete(entry.workId);
-    entry.resolveSettlement();
-  };
-
-  const requestStop = (entry) => {
-    if (entry.stopPromise !== null) return;
-    entry.stopPromise = Promise.resolve()
-      .then(entry.stop)
-      .catch((error) => {
-        shutdownErrors.push(contextualError(
-          `Failed to stop mutable PostgreSQL integration work during terminal shutdown. phase=${entry.phase}`,
-          error,
-        ));
-      })
-      .finally(() => {
-        entry.stopSettled = true;
-        finishEntry(entry);
-      });
-  };
-
-  const awaitTrackedOperation = (entry, operation, operationPhase) => {
-    const rawOperation = Promise.resolve(operation);
-    const operationState = {
-      callerInterrupted: false,
-      callerSettled: false,
-      operationPhase,
-      rawOperation,
-    };
-    entry.rawOperations.add(operationState);
-    rawOperation.then(
-      () => {
-        entry.rawOperations.delete(operationState);
-        finishEntry(entry);
-      },
-      (error) => {
-        if (operationState.callerInterrupted) {
-          shutdownErrors.push(contextualError(
-            `Mutable PostgreSQL integration operation failed after terminal shutdown interrupted its caller. phase=${entry.phase} operationPhase=${operationPhase}`,
-            error,
-          ));
-        }
-        entry.rawOperations.delete(operationState);
-        finishEntry(entry);
-      },
-    );
-
-    const operationForCaller = rawOperation.then(
-      (result) => {
-        operationState.callerSettled = true;
-        return result;
-      },
-      (error) => {
-        operationState.callerSettled = true;
-        throw error;
-      },
-    );
-    const shutdownForCaller = shutdown.then((failure) => {
-      if (!operationState.callerSettled) {
-        operationState.callerInterrupted = true;
-        operationState.callerSettled = true;
-      }
-      throw contextualError(
-        `PostgreSQL integration work was interrupted by terminal shutdown. phase=${operationPhase}`,
-        failure,
-      );
-    });
-    return Promise.race([operationForCaller, shutdownForCaller]);
-  };
-
-  return Object.freeze({
-    assertCanStart(phase) {
-      if (firstShutdownFailure === null) return;
-      throw contextualError(
-        `Refusing to start mutable PostgreSQL integration work after terminal shutdown began. phase=${phase}`,
-        firstShutdownFailure,
-      );
-    },
-    async awaitOperation(operation, phase) {
-      if (firstShutdownFailure !== null) {
-        throw contextualError(
-          `Refusing to await mutable PostgreSQL integration work after terminal shutdown began. phase=${phase}`,
-          firstShutdownFailure,
-        );
-      }
-      return Promise.race([
-        Promise.resolve(operation),
-        shutdown.then((failure) => {
-          throw contextualError(
-            `PostgreSQL integration work was interrupted by terminal shutdown. phase=${phase}`,
-            failure,
-          );
-        }),
-      ]);
-    },
-    errors() {
-      return [...shutdownErrors];
-    },
-    register(phase, stop) {
-      let resolveSettlement;
-      const settlement = new Promise((resolveRegisteredWork) => {
-        resolveSettlement = resolveRegisteredWork;
-      });
-      const workId = nextWorkId;
-      nextWorkId += 1;
-      const entry = {
-        completed: false,
-        phase,
-        rawOperations: new Set(),
-        resolveSettlement,
-        settlement,
-        stop,
-        stopPromise: null,
-        stopSettled: false,
-        workId,
-      };
-      activeWork.set(workId, entry);
-      if (firstShutdownFailure !== null) requestStop(entry);
-
-      return Object.freeze({
-        assertCanStart() {
-          if (firstShutdownFailure === null) return;
-          throw contextualError(
-            `Refusing to start mutable PostgreSQL integration work after terminal shutdown began. phase=${phase}`,
-            firstShutdownFailure,
-          );
-        },
-        awaitOperation(operation, operationPhase) {
-          return awaitTrackedOperation(
-            entry,
-            operation,
-            operationPhase,
-          );
-        },
-        complete() {
-          if (entry.completed) return;
-          entry.completed = true;
-          finishEntry(entry);
-        },
-      });
-    },
-    stopAll(failure) {
-      if (firstShutdownFailure === null) {
-        firstShutdownFailure = failure;
-        resolveShutdown(failure);
-      }
-      for (const entry of activeWork.values()) {
-        requestStop(entry);
-      }
-    },
-    async waitForStopped(cleanupDeadline) {
-      if (activeWork.size === 0) return;
-      while (activeWork.size > 0) {
-        const entries = [...activeWork.values()];
-        if (firstShutdownFailure !== null) {
-          for (const entry of entries) requestStop(entry);
-        }
-        const remainingMilliseconds = cleanupRemainingMilliseconds(
-          cleanupDeadline,
-          `mutable work shutdown activePhases=${JSON.stringify(entries.map((entry) => entry.phase))}`,
-        );
-        await Promise.race([
-          Promise.all(entries.map((entry) => entry.settlement)),
-          delay(Math.min(
-            mutableWorkShutdownPollMilliseconds,
-            remainingMilliseconds,
-          )),
-        ]);
-      }
-    },
-  });
-}
-
-async function withPostgresClient(
-  clientOptions,
-  phase,
-  callback,
-  mutableWorkSupervisor,
-) {
-  const supervisedClient = createSupervisedPostgresClient(
-    clientOptions,
-    phase,
-    null,
-  );
-  const { client } = supervisedClient;
-  const mutableWork = mutableWorkSupervisor.register(
-    phase,
-    () => supervisedClient.abort(),
-  );
-  let result;
-  let primaryError = null;
-  try {
-    mutableWork.assertCanStart();
-    await mutableWork.awaitOperation(
-      client.connect(),
-      `${phase} connection establishment`,
-    );
-    mutableWork.assertCanStart();
-    await mutableWork.awaitOperation(
-      requirePostgresSessionContract(
-        client,
-        clientOptions.database,
-        `${phase} session contract verification`,
-      ),
-      `${phase} session contract verification`,
-    );
-    mutableWork.assertCanStart();
-    result = await mutableWork.awaitOperation(
-      callback(client),
-      phase,
-    );
-  } catch (error) {
-    primaryError = contextualError(
-      `PostgreSQL integration client operation failed. phase=${phase}`,
-      error,
-    );
-  }
-
-  let closeError = null;
-  try {
-    await mutableWork.awaitOperation(
-      supervisedClient.close(),
-      `${phase} client close`,
-    );
-  } catch (error) {
-    closeError = contextualError(
-      `PostgreSQL integration client close failed. phase=${phase}`,
-      error,
-    );
-  }
-  const backgroundFailure = supervisedClient.backgroundFailure();
-  supervisedClient.detach();
-  mutableWork.complete();
-  const failure = combineErrors(
-    [
-      ...(primaryError === null ? [] : [primaryError]),
-      ...(backgroundFailure === null ? [] : [backgroundFailure]),
-      ...(closeError === null ? [] : [closeError]),
-    ],
-    `PostgreSQL integration operation and client close failed. phase=${phase}`,
-  );
-  if (failure !== null) throw failure;
-  return result;
-}
-
-async function acquireLifecycleLock(adminClient) {
-  const result = await adminClient.query(
-    `SELECT pg_catalog.pg_try_advisory_lock($1::integer, $2::integer) AS acquired`,
-    lifecycleLockKeys,
-  );
-  if (result.rows[0]?.acquired !== true) {
-    throw new Error(
-      "Another PostgreSQL integration runner already owns the cluster lifecycle lock. Use a separate isolated PostgreSQL 18 cluster or wait for that runner to finish.",
-    );
-  }
-}
-
-async function reacquireLifecycleLockForCleanup(
-  cleanupSession,
-  cleanupDeadline,
-) {
-  while (true) {
-    const result = await runCleanupQuery(
-      cleanupSession,
-      cleanupDeadline,
-      "lifecycle lock reacquisition",
-      `SELECT pg_catalog.pg_try_advisory_lock($1::integer, $2::integer) AS acquired`,
-      lifecycleLockKeys,
-    );
-    if (result.rows[0]?.acquired === true) return;
-    const remainingMilliseconds = cleanupRemainingMilliseconds(
-      cleanupDeadline,
-      "lifecycle lock reacquisition wait",
-    );
-    await delay(Math.min(
-      lifecycleRecoveryPollMilliseconds,
-      remainingMilliseconds,
-    ));
-  }
-}
-
-async function terminateDatabaseSessionsByOid(
-  cleanupSession,
-  cleanupDeadline,
-  databaseOid,
-  phase,
-) {
-  const terminationResult = await runCleanupQuery(
-    cleanupSession,
-    cleanupDeadline,
-    `${phase} termination signals`,
-    `SELECT
-       pid,
-       pg_catalog.pg_terminate_backend(
-         pid,
-         0
-       ) AS terminated
-     FROM pg_catalog.pg_stat_activity
-     WHERE datid = $1::oid AND pid <> pg_catalog.pg_backend_pid()
-     ORDER BY pid`,
-    [databaseOid],
-  );
-  const unsignaledSessions = terminationResult.rows
-    .filter((row) => row.terminated !== true)
-    .map((row) => Object.freeze({
-      pid: row.pid,
-      signalAccepted: row.terminated,
-    }));
-
-  let remainingPids = [];
-  while (true) {
-    const sessionsResult = await runCleanupQuery(
-      cleanupSession,
-      cleanupDeadline,
-      `${phase} termination polling`,
-      `SELECT pid
-       FROM pg_catalog.pg_stat_activity
-       WHERE datid = $1::oid AND pid <> pg_catalog.pg_backend_pid()
-       ORDER BY pid`,
-      [databaseOid],
-    );
-    remainingPids = sessionsResult.rows.map((row) => row.pid);
-    if (remainingPids.length === 0) break;
-    const remainingMilliseconds = Math.floor(
-      cleanupDeadline.expiresAt - performance.now(),
-    );
-    if (remainingMilliseconds <= cleanupClientTeardownReserveMilliseconds) {
-      break;
-    }
-    await delay(Math.min(
-      databaseTerminationPollMilliseconds,
-      remainingMilliseconds - cleanupClientTeardownReserveMilliseconds,
-    ));
-  }
-
-  if (unsignaledSessions.length > 0 || remainingPids.length > 0) {
-    throw new Error(
-      `PostgreSQL could not terminate every session on the exact runner-owned database OID within one shared deadline. databaseOid=${databaseOid} phase=${phase} timeoutMilliseconds=${cleanupDeadline.timeoutMilliseconds} unsignaledSessions=${JSON.stringify(unsignaledSessions)} remainingPids=${JSON.stringify(remainingPids)}`,
-    );
-  }
-}
-
-function getDatabaseOwnershipEvidence(boundaryState) {
-  if (
-    boundaryState.pendingDatabaseOid !== null
-    && boundaryState.ownedDatabaseOid !== null
-  ) {
-    throw new Error(
-      `PostgreSQL integration database ownership state contains both pending and promoted identities. database=${disposableDatabaseName} pendingOid=${boundaryState.pendingDatabaseOid} ownedOid=${boundaryState.ownedDatabaseOid}`,
-    );
-  }
-  if (boundaryState.ownedDatabaseOid !== null) {
-    return Object.freeze({
-      databaseOid: boundaryState.ownedDatabaseOid,
-      ownershipState: "owned",
-    });
-  }
-  if (boundaryState.pendingDatabaseOid !== null) {
-    return Object.freeze({
-      databaseOid: boundaryState.pendingDatabaseOid,
-      ownershipState: "pending",
-    });
-  }
-  return null;
-}
-
-function classifyDatabaseIdentityRows(rows, expectedDatabaseOid) {
-  const namedDatabase = rows.find(
-    (row) => row.datname === disposableDatabaseName,
-  );
-  const oidDatabase = rows.find(
-    (row) => row.database_oid === expectedDatabaseOid,
-  );
-  const nameMatch = namedDatabase === undefined
-    ? null
-    : Object.freeze({
-      databaseName: namedDatabase.datname,
-      databaseOid: namedDatabase.database_oid,
-    });
-  const oidMatch = oidDatabase === undefined
-    ? null
-    : Object.freeze({
-      databaseName: oidDatabase.datname,
-      databaseOid: oidDatabase.database_oid,
-    });
-
-  let classification;
-  if (nameMatch === null && oidMatch === null) {
-    classification = "absent";
-  } else if (
-    nameMatch?.databaseOid === expectedDatabaseOid
-    && oidMatch?.databaseName === disposableDatabaseName
-  ) {
-    classification = "exact";
-  } else if (nameMatch !== null && oidMatch !== null) {
-    classification = "name-and-oid-conflict";
-  } else if (nameMatch !== null) {
-    classification = "name-conflict";
-  } else {
-    classification = "oid-conflict";
-  }
-
-  return Object.freeze({
-    classification,
-    expectedDatabaseOid,
-    nameMatch,
-    oidMatch,
-  });
-}
-
-async function inspectDatabaseIdentity(adminClient, expectedDatabaseOid) {
-  const result = await adminClient.query(
-    `SELECT datname, oid::text AS database_oid
-     FROM pg_catalog.pg_database
-     WHERE datname = $1 OR oid = $2::oid
-     ORDER BY datname`,
-    [disposableDatabaseName, expectedDatabaseOid],
-  );
-  return classifyDatabaseIdentityRows(result.rows, expectedDatabaseOid);
-}
-
-async function inspectDatabaseIdentityForCleanup(
-  cleanupSession,
-  cleanupDeadline,
-  expectedDatabaseOid,
-  phase,
-) {
-  const result = await runCleanupQuery(
-    cleanupSession,
-    cleanupDeadline,
-    phase,
-    `SELECT datname, oid::text AS database_oid
-     FROM pg_catalog.pg_database
-     WHERE datname = $1 OR oid = $2::oid
-     ORDER BY datname`,
-    [disposableDatabaseName, expectedDatabaseOid],
-  );
-  return classifyDatabaseIdentityRows(result.rows, expectedDatabaseOid);
-}
-
-function databaseIdentityInterferenceError(identity, phase) {
-  return new Error(
-    `Refusing to claim or clean a PostgreSQL database after name/OID ownership interference. phase=${phase} expectedDatabase=${disposableDatabaseName} expectedOid=${identity.expectedDatabaseOid} classification=${identity.classification} nameMatchDatabase=${identity.nameMatch?.databaseName ?? "missing"} nameMatchOid=${identity.nameMatch?.databaseOid ?? "missing"} oidMatchDatabase=${identity.oidMatch?.databaseName ?? "missing"} oidMatchOid=${identity.oidMatch?.databaseOid ?? "missing"}`,
-  );
-}
-
-async function inspectOwnedRoleIdentities(
-  cleanupSession,
-  cleanupDeadline,
-  boundaryState,
-  phase,
-) {
-  const ownedRoles = [...boundaryState.ownedRoleOids.entries()];
-  if (ownedRoles.length === 0) return Object.freeze([]);
-
-  const result = await runCleanupQuery(
-    cleanupSession,
-    cleanupDeadline,
-    phase,
-    `SELECT rolname, oid::text AS role_oid
-     FROM pg_catalog.pg_roles
-     WHERE rolname = ANY($1::text[]) OR oid = ANY($2::oid[])
-     ORDER BY rolname`,
-    [
-      ownedRoles.map(([roleName]) => roleName),
-      ownedRoles.map(([, roleOid]) => roleOid),
-    ],
-  );
-  return Object.freeze(ownedRoles.map(([roleName, expectedRoleOid]) => {
-    const namedRole = result.rows.find((row) => row.rolname === roleName);
-    const oidRole = result.rows.find(
-      (row) => row.role_oid === expectedRoleOid,
-    );
-    const nameMatch = namedRole === undefined
-      ? null
-      : Object.freeze({
-        roleName: namedRole.rolname,
-        roleOid: namedRole.role_oid,
-      });
-    const oidMatch = oidRole === undefined
-      ? null
-      : Object.freeze({
-        roleName: oidRole.rolname,
-        roleOid: oidRole.role_oid,
-      });
-
-    let classification;
-    if (nameMatch === null && oidMatch === null) {
-      classification = "absent";
-    } else if (
-      nameMatch?.roleOid === expectedRoleOid
-      && oidMatch?.roleName === roleName
-    ) {
-      classification = "exact";
-    } else if (nameMatch !== null && oidMatch !== null) {
-      classification = "name-and-oid-conflict";
-    } else if (nameMatch !== null) {
-      classification = "name-conflict";
-    } else {
-      classification = "oid-conflict";
-    }
-
-    return Object.freeze({
-      classification,
-      expectedRoleOid,
-      nameMatch,
-      oidMatch,
-      roleName,
-    });
-  }));
-}
-
-function roleIdentityInterferenceError(identity, phase) {
-  return new Error(
-    `Refusing to clean a PostgreSQL role after name/OID ownership interference. phase=${phase} expectedRole=${identity.roleName} expectedOid=${identity.expectedRoleOid} classification=${identity.classification} nameMatchRole=${identity.nameMatch?.roleName ?? "missing"} nameMatchOid=${identity.nameMatch?.roleOid ?? "missing"} oidMatchRole=${identity.oidMatch?.roleName ?? "missing"} oidMatchOid=${identity.oidMatch?.roleOid ?? "missing"}`,
-  );
-}
-
-async function inspectBoundaryCleanupIdentities(
-  cleanupSession,
-  cleanupDeadline,
-  boundaryState,
-) {
-  const databaseOwnership = getDatabaseOwnershipEvidence(boundaryState);
-  const databaseIdentity = databaseOwnership === null
-    ? null
-    : await inspectDatabaseIdentityForCleanup(
-      cleanupSession,
-      cleanupDeadline,
-      databaseOwnership.databaseOid,
-      "whole-boundary database identity classification",
-    );
-  const roleIdentities = await inspectOwnedRoleIdentities(
-    cleanupSession,
-    cleanupDeadline,
-    boundaryState,
-    "whole-boundary role identity classification",
-  );
-  return Object.freeze({
-    databaseIdentity,
-    databaseOwnership,
-    roleIdentities,
-  });
-}
-
-function cleanupIdentityConflictErrors(cleanupIdentities) {
-  const conflicts = [];
-  if (
-    cleanupIdentities.databaseIdentity !== null
-    && cleanupIdentities.databaseIdentity.classification !== "exact"
-    && cleanupIdentities.databaseIdentity.classification !== "absent"
-  ) {
-    conflicts.push(databaseIdentityInterferenceError(
-      cleanupIdentities.databaseIdentity,
-      "whole-boundary cleanup classification",
-    ));
-  }
-  for (const roleIdentity of cleanupIdentities.roleIdentities) {
-    if (
-      roleIdentity.classification === "exact"
-      || roleIdentity.classification === "absent"
-    ) {
-      continue;
-    }
-    conflicts.push(roleIdentityInterferenceError(
-      roleIdentity,
-      "whole-boundary cleanup classification",
-    ));
-  }
-  return conflicts;
-}
 
 async function releaseLifecycleLock(cleanupSession, cleanupDeadline) {
   const result = await runCleanupQuery(
